[BDE] Engine owns 100% of KV; consolidate kv_cache into 4 modules
Two related changes to the BDE KV stack (both verified bit-identical):

Engine-owned KV (removes the last model-local fallbacks):
- Self-attn: build_window_slots returns an empty index for a 0-block window,
  so prefill-start gathers a zero-length window (the concat-identity the
  model expects) straight from the pool. Dropped BDEKVState.get_kv_caches'
  `_committed==0` fallback + its `fallback` param.
- Cross-attn: get_cross_kv_caches now asserts the cross pool is populated
  (raises) instead of falling back to model-local lazy projection.
- pipeline _kv_get/_kv_get_cross drop the model-local lambdas; _kv_create is
  skipped under BDE (pool owns allocation). Model + non-BDE path untouched.
- Verified: KV-off vs KV-on on DreamZero-DROID is bit-identical
  (PSNR=inf, SSIM~1.0, max_abs_diff=0.0, LPIPS=0.0).

Consolidate kv_cache 7 -> 4 files (pure move, public API unchanged):
- config.py (knob) · paged.py (engine-generic slot-mapping + pool I/O +
  chunk-window eviction spec) · manager.py (BDEKVCache + BDERequestAdapter +
  pool builders) · state.py (the model-facing BDEKVState bridge).
- Removed adapter.py / chunk_window.py / gather.py / pool.py / slot_mapping.py;
  __init__ re-exports the same names. Mirrors vLLM's granularity and keeps the
  generic-vs-bridge seam the Cosmos port reuses.

Tests: tests/bde 79 passed; full bde-or-dreamzero suite green on GPU.
