+++ vllm_omni/experimental/bde/kv_cache/__init__.py
@@ -4,21 +4,27 @@
 Thin glue over vLLM's paged KV stack (``KVCacheManager`` / ``BlockPool`` /
 ``SlidingWindowManager``), used by the Block Diffusion Engine to manage KV for
 AR-diffusion models. See ``BDE_doc/dreamzero_kv_phase1_plan.md``.
+
+Layout: ``config`` (public knob) · ``paged`` (engine-generic paging mechanics +
+chunk-window eviction spec) · ``manager`` (the BDEKVCache orchestrator + its
+request adapter / pool builders) · ``state`` (the model-facing BDEKVState bridge).
 """
 
-from vllm_omni.experimental.bde.kv_cache.adapter import BDERequestAdapter
-from vllm_omni.experimental.bde.kv_cache.chunk_window import ChunkWindowManager, ChunkWindowSpec
 from vllm_omni.experimental.bde.kv_cache.config import BDEKVConfig
-from vllm_omni.experimental.bde.kv_cache.gather import (
-    allocate_kv_pool,
-    pool_gather_window,
-    pool_write_chunk,
+from vllm_omni.experimental.bde.kv_cache.manager import (
+    BDEKVCache,
+    BDERequestAdapter,
+    build_kv_manager,
+    compute_num_blocks,
 )
-from vllm_omni.experimental.bde.kv_cache.manager import BDEKVCache
-from vllm_omni.experimental.bde.kv_cache.pool import build_kv_manager, compute_num_blocks
-from vllm_omni.experimental.bde.kv_cache.slot_mapping import (
+from vllm_omni.experimental.bde.kv_cache.paged import (
+    ChunkWindowManager,
+    ChunkWindowSpec,
+    allocate_kv_pool,
     chunk_slot_mapping,
     compute_slot_mapping,
+    pool_gather_window,
+    pool_write_chunk,
     resident_block_ids,
 )
