+++ vllm_omni/experimental/bde/runner.py
@@ -25,7 +25,7 @@ from vllm_omni.diffusion.models.dreamzero.pipeline_dreamzero import MAX_DREAMZER
 from vllm_omni.diffusion.request import OmniDiffusionRequest
 from vllm_omni.diffusion.worker.diffusion_model_runner import DiffusionModelRunner
 from vllm_omni.experimental.bde.kv_cache.config import BDEKVConfig
-from vllm_omni.experimental.bde.kv_cache.gather import BDEKVState
+from vllm_omni.experimental.bde.kv_cache.state import BDEKVState
 from vllm_omni.experimental.bde.kv_cache.manager import BDEKVCache
 
 logger = init_logger(__name__)
